Add <meta name="viewport"> tag and javascript hack
The viewport meta tag is needed to make pages scale properly on mobile
devices. This is not a standard, but a convention. And unfortunately, the
convention is stupid and doesn't provide a way to specify a minimum
viewport width. You can specify either a fixed width that does not scale
to larger screens, or "device-width" which uses the device width but makes
no concession for layouts that happen to be wider than that; the browser
will load the page zoomed in and possibly not even allow the user to zoom
out. It's all bullshit really and the only way to have a viewport of at
least a certain size but still scalable to larger screens is to hack it
around with Javascript to set a fixed-width viewport equal to screen.width.