Create blog layout
Add twocols template for a two-column flexbox layout, and extend a blogbase
template from it. The blogbase template has a tree-based navigation sidebar on
the left column and blog content on the right; this will be the base layout for
all blog pages. Currently, only the blog index is implemented.